# Align StoredTableNodes across multiple LQPs (include all subqueries)