--- a/src/lib/logical_query_plan/lqp_utils.hpp
+++ b/src/lib/logical_query_plan/lqp_utils.hpp
@@ -19,15 +19,50 @@ class LQPSubqueryExpression;
 enum class LQPInputSide;
 
 using LQPMismatch = std::pair<std::shared_ptr<const AbstractLQPNode>, std::shared_ptr<const AbstractLQPNode>>;
+
+/**
+ * Data structure that maps LQPs to (multiple) subquery expressions that reference them.
+ *
+ * PURPOSE:
+ *  Mainly used by optimizer rules to optimize subquery subplans more efficiently. In concrete, it helps
+ *  to optimize subquery subplans ONLY ONCE, although being referenced by a list of subquery expressions.
+ *
+ * WHY WEAK POINTERS FOR SUBQUERY EXPRESSIONS?
+ *  Referenced LQPs and subquery expressions might be objects to change after creating this data structure. Depending
+ *  on the order of optimization steps, we could end up with a scenario as follows:
+ *
+ *      [ProjectionNodeRoot(..., SubqueryExpressionA)]
+ *                                        \
+ *                                         \ references
+ *                                          \
+ *                                        [ProjectionNodeA(..., SubqueryExpressionB)]
+ *                                                                       \
+ *                                                                        \ references
+ *                                                                         \
+ *                                                                        [ProjectionNodeB(...)]
+ *
+ *        (1) collect_subquery_expressions_by_lqp(ProjectionNodeRoot)
+ *                =>  returns { [ ProjectionNodeA, { SubqueryExpressionA } ],
+ *                              [ ProjectionNodeB, { SubqueryExpressionB } ] }
+ *
+ *        (2) OptimizerRuleXY is applied to ProjectionNodeA
+ *                => As a result, SubqueryExpressionB gets replaced / removed from ProjectionNodeA
+ *
+ *        (3) OptimizerRuleXY is applied to ProjectionNodeB
+ *            !!! WASTED OPTIMIZATION TIME because SubqueryExpressionB and ProjectionNodeB are no longer being used,
+ *                                                                                                thanks to step (2) !!!
+ *
+ *  With weak pointers, we are FORCED TO SKIP STEP (3) because SubqueryExpressionB and its corresponding LQP have
+ *  already been removed from memory after STEP (2).
+ *
+ *  However, this optimization does not cover all cases as it is dependent on the execution order. For
+ *  example, when swapping steps (2) and (3), we cannot easily skip an optimization step.
+ */
 using SubqueryExpressionsByLQP =
     std::vector<std::pair<std::shared_ptr<AbstractLQPNode>, std::vector<std::weak_ptr<LQPSubqueryExpression>>>>;
 
 /**
- * TODO(Julian) Doc
- * Traverses @param node with all its (nested) subquery expressions to identify unique LQPs and the (multiple)
- * SubqueryExpressions referencing each of these unique LQPs.
- * @param node
- * @return
+ * Returns unique LQPs from (nested) subquery expressions of @param node.
  */
 SubqueryExpressionsByLQP collect_subquery_expressions_by_lqp(const std::shared_ptr<AbstractLQPNode>& node);