--- a/src/lib/optimizer/optimizer.cpp
+++ b/src/lib/optimizer/optimizer.cpp
@@ -125,7 +125,7 @@ std::shared_ptr<AbstractLQPNode> Optimizer::optimize(
 
   for (const auto& rule : _rules) {
     Timer rule_timer{};
-    rule->apply_to(root_node);
+    rule->apply_to_plan(root_node);
     auto rule_duration = rule_timer.lap();
 
     if (rule_durations) {