--- a/src/lib/optimizer/strategy/abstract_rule.cpp
+++ b/src/lib/optimizer/strategy/abstract_rule.cpp
@@ -34,6 +34,7 @@ void AbstractRule::apply(const std::shared_ptr<AbstractLQPNode>& lqp_root) const
   for (const auto& [lqp, subquery_expressions] : subquery_expressions_by_lqp) {
     // (2.1) Prepare
     const auto local_lqp_root = LogicalPlanRootNode::make(lqp);
+    std::cout << "Subquery LQP \n\n" << *lqp << "\n\n" << std::endl;
     // (2.2) Optimize subquery LQP
     _apply_to(local_lqp_root);
     // (2.3) Assign optimized LQP to all corresponding SubqueryExpressions