… src/lib/optimizer/strategy/abstract_rule.hpp
… src/lib/optimizer/strategy/between_composition_rule.cpp
… src/lib/optimizer/strategy/between_composition_rule.hpp
… src/lib/optimizer/strategy/chunk_pruning_rule.cpp
… src/lib/optimizer/strategy/chunk_pruning_rule.hpp
… src/lib/optimizer/strategy/column_pruning_rule.cpp
… src/lib/optimizer/strategy/column_pruning_rule.hpp
… src/lib/optimizer/strategy/dependent_group_by_reduction_rule.cpp
… src/lib/optimizer/strategy/dependent_group_by_reduction_rule.hpp
… src/lib/optimizer/strategy/expression_reduction_rule.cpp
… src/lib/optimizer/strategy/expression_reduction_rule.hpp
… src/lib/optimizer/strategy/in_expression_rewrite_rule.cpp
… src/lib/optimizer/strategy/in_expression_rewrite_rule.hpp
… src/lib/optimizer/strategy/index_scan_rule.cpp
… src/lib/optimizer/strategy/index_scan_rule.hpp
… src/lib/optimizer/strategy/join_ordering_rule.cpp
… src/lib/optimizer/strategy/join_ordering_rule.hpp
… src/lib/optimizer/strategy/join_predicate_ordering_rule.cpp
… src/lib/optimizer/strategy/join_predicate_ordering_rule.hpp
… src/lib/optimizer/strategy/predicate_merge_rule.cpp
… src/lib/optimizer/strategy/predicate_merge_rule.hpp
… src/lib/optimizer/strategy/predicate_placement_rule.cpp
… src/lib/optimizer/strategy/predicate_placement_rule.hpp
… src/lib/optimizer/strategy/predicate_reordering_rule.cpp
… src/lib/optimizer/strategy/predicate_reordering_rule.hpp
… src/lib/optimizer/strategy/predicate_split_up_rule.cpp
… src/lib/optimizer/strategy/predicate_split_up_rule.hpp
… src/lib/optimizer/strategy/semi_join_reduction_rule.cpp
… src/lib/optimizer/strategy/semi_join_reduction_rule.hpp
… src/lib/optimizer/strategy/stored_table_column_alignment_rule.cpp