Partial changes from #2254 to simplify the review process.

**(1)** Logic regarding subquery handling is moved from the `Optimizer` to `AbstractRule`. With this change, rules can override the default subquery handling logic.

- `Optimizer::_apply_rule` becomes `AbstractRule::apply_to_plan` (virtual)
- `AbstractRule::apply_to` becomes `AbstractRule::_apply_to_plan_without_subqueries` (pure virtual)

**(2)** The StoredTableAlignmentRule utilizes the changes from (1) and overrides the subquery handling behavior for improved alignment.

* **Before:** Root and subquery LQPs were optimized individually. Therefore, there was no cross-LQP alignment of StoredTableNodes.

* **After:** Root and subquery LQPs are optimized together. Hence, StoredTableNodes are aligned across all LQPs.

---

No performance boost yet. We need to have subplan deduplication for subqueries, which comes with #2254.

**System**
<details>
<summary>pella - click to expand</summary>

| property | value |
| -- | -- |
| Hostname | pella |
| CPU | Intel(R) Xeon(R) CPU E7- 8870  @ 2.40GHz |
| Memory | 1.4TB |
| numactl | nodebind: 2  |
| numactl | membind: 2  |
</details>

**Commit Info and Build Time**
| commit | date | message | build time |
| -- | -- | -- | -- |
| a7e36e433 | 19.11.2020 11:16 | Fix sorted aggregation micro benchmarks (#2265) | real 527.38 user 4094.76 sys 173.50|
| 33de60b53 | 24.11.2020 17:48 | Move recursion logic to collect_subquery_expressions_by_lqp() | real 529.60 user 4169.08 sys 176.50|


**hyriseBenchmarkTPCH - single-threaded**
<details>
<summary>
Sum of avg. item runtimes: -0%
 ||
Geometric mean of throughput changes: +0%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_a7e36e433e40422ba07f09f0b58088d26703c64f_st.json | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_33de60b53bd65430ec4796dca434285058951992_st.json |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | a7e36e433e40422ba07f09f0b58088d26703c64f-dirty                                                                                             | 33de60b53bd65430ec4796dca434285058951992-dirty                                                                                             |
 |  benchmark_mode          | Ordered                                                                                                                                    | Ordered                                                                                                                                    |
 |  build_type              | release                                                                                                                                    | release                                                                                                                                    |
 |  chunk_size              | 65535                                                                                                                                      | 65535                                                                                                                                      |
 |  clients                 | 1                                                                                                                                          | 1                                                                                                                                          |
 |  compiler                | gcc 9.2                                                                                                                                    | gcc 9.2                                                                                                                                    |
 |  cores                   | 0                                                                                                                                          | 0                                                                                                                                          |
 |  date                    | 2020-11-24 18:26:48                                                                                                                        | 2020-11-24 21:33:52                                                                                                                        |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                    | {'default': {'encoding': 'Dictionary'}}                                                                                                    |
 |  indexes                 | False                                                                                                                                      | False                                                                                                                                      |
 |  max_duration            | 60000000000                                                                                                                                | 60000000000                                                                                                                                |
 |  max_runs                | -1                                                                                                                                         | -1                                                                                                                                         |
 |  scale_factor            | 1.0                                                                                                                                        | 1.0                                                                                                                                        |
 |  time_unit               | ns                                                                                                                                         | ns                                                                                                                                         |
 |  use_prepared_statements | False                                                                                                                                      | False                                                                                                                                      |
 |  using_scheduler         | False                                                                                                                                      | False                                                                                                                                      |
 |  verify                  | False                                                                                                                                      | False                                                                                                                                      |
 |  warmup_duration         | 0                                                                                                                                          | 0                                                                                                                                          |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||   1480.8 |  1464.9 |   -1%  ||     0.68 |     0.68 |   +1%  |  0.4933 |
 | TPC-H 02 ||      9.8 |     9.8 |   -0%  ||   102.31 |   102.37 |   +0%  |  0.9561 |
 | TPC-H 03 ||    156.2 |   156.1 |   -0%  ||     6.40 |     6.41 |   +0%  |  0.8024 |
 | TPC-H 04 ||    139.7 |   139.9 |   +0%  ||     7.16 |     7.15 |   -0%  |  0.2565 |
 | TPC-H 05 ||    428.0 |   438.0 |   +2%  ||     2.34 |     2.28 |   -2%  |  0.0000 |
 | TPC-H 06 ||      8.3 |     8.3 |   -0%  ||   120.74 |   121.01 |   +0%  |  0.3990 |
 | TPC-H 07 ||    115.3 |   115.8 |   +1%  ||     8.67 |     8.63 |   -1%  |  0.6476 |
 | TPC-H 08 ||    122.5 |   120.8 |   -1%  ||     8.17 |     8.27 |   +1%  |  0.0233 |
 | TPC-H 09 ||    876.5 |   871.2 |   -1%  ||     1.14 |     1.15 |   +1%  |  0.1863 |
 | TPC-H 10 ||    206.5 |   206.5 |   +0%  ||     4.84 |     4.84 |   -0%  |  0.8322 |
 | TPC-H 11 ||     13.5 |    13.5 |   +0%  ||    74.06 |    73.97 |   -0%  |  0.3353 |
 | TPC-H 12 ||     85.6 |    85.5 |   -0%  ||    11.68 |    11.69 |   +0%  |  0.6972 |
 | TPC-H 13 ||    656.6 |   657.5 |   +0%  ||     1.52 |     1.52 |   -0%  |  0.5059 |
 | TPC-H 14 ||     38.1 |    37.6 |   -1%  ||    26.26 |    26.57 |   +1%  |  0.0000 |
 | TPC-H 15 ||     15.1 |    15.1 |   +0%  ||    66.19 |    66.10 |   -0%  |  0.1277 |
 | TPC-H 16 ||    128.3 |   127.6 |   -1%  ||     7.79 |     7.84 |   +1%  |  0.0000 |
 | TPC-H 17 ||     43.3 |    43.0 |   -1%  ||    23.09 |    23.24 |   +1%  |  0.0168 |
 | TPC-H 18 ||    927.3 |   926.9 |   -0%  ||     1.08 |     1.08 |   +0%  |  0.8311 |
 | TPC-H 19 ||     65.5 |    65.4 |   -0%  ||    15.27 |    15.29 |   +0%  |  0.3221 |
 | TPC-H 20 ||     28.5 |    28.5 |   -0%  ||    35.05 |    35.08 |   +0%  |  0.8125 |
 | TPC-H 21 ||    641.3 |   642.2 |   +0%  ||     1.56 |     1.56 |   -0%  |  0.7839 |
 | TPC-H 22 ||     70.8 |    70.4 |   -0%  ||    14.13 |    14.20 |   +0%  |  0.0011 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||   6257.2 |  6244.7 |   -0%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   +0%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```
</details>


**hyriseBenchmarkTPCH - single-threaded, SF 0.01**
<details>
<summary>
Sum of avg. item runtimes: -1%
 ||
Geometric mean of throughput changes: +0%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_a7e36e433e40422ba07f09f0b58088d26703c64f_st_s01.json | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_33de60b53bd65430ec4796dca434285058951992_st_s01.json |
 +--------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | a7e36e433e40422ba07f09f0b58088d26703c64f-dirty                                                                                                 | 33de60b53bd65430ec4796dca434285058951992-dirty                                                                                                 |
 |  benchmark_mode          | Ordered                                                                                                                                        | Ordered                                                                                                                                        |
 |  build_type              | release                                                                                                                                        | release                                                                                                                                        |
 |  chunk_size              | 65535                                                                                                                                          | 65535                                                                                                                                          |
 |  clients                 | 1                                                                                                                                              | 1                                                                                                                                              |
 |  compiler                | gcc 9.2                                                                                                                                        | gcc 9.2                                                                                                                                        |
 |  cores                   | 0                                                                                                                                              | 0                                                                                                                                              |
 |  date                    | 2020-11-24 18:49:10                                                                                                                            | 2020-11-24 21:56:12                                                                                                                            |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                        | {'default': {'encoding': 'Dictionary'}}                                                                                                        |
 |  indexes                 | False                                                                                                                                          | False                                                                                                                                          |
 |  max_duration            | 60000000000                                                                                                                                    | 60000000000                                                                                                                                    |
 |  max_runs                | -1                                                                                                                                             | -1                                                                                                                                             |
 |  scale_factor            | 0.009999999776482582                                                                                                                           | 0.009999999776482582                                                                                                                           |
 |  time_unit               | ns                                                                                                                                             | ns                                                                                                                                             |
 |  use_prepared_statements | False                                                                                                                                          | False                                                                                                                                          |
 |  using_scheduler         | False                                                                                                                                          | False                                                                                                                                          |
 |  verify                  | False                                                                                                                                          | False                                                                                                                                          |
 |  warmup_duration         | 0                                                                                                                                              | 0                                                                                                                                              |
 +--------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||     13.5 |    13.3 |   -1%  ||    74.25 |    74.91 |   +1%  |  0.0000 |
 | TPC-H 02 ||      1.7 |     1.7 |   +0%  ||   580.28 |   579.66 |   -0%  |  0.8816 |
 | TPC-H 03 ||      1.5 |     1.5 |   -0%  ||   649.08 |   649.46 |   +0%  |  0.5048 |
 | TPC-H 04 ||      1.0 |     1.0 |   -0%  ||   992.28 |   996.10 |   +0%  |  0.0000 |
 | TPC-H 05 ||      2.5 |     2.5 |   +0%  ||   399.02 |   398.79 |   -0%  |  0.8016 |
 | TPC-H 06 ||      0.3 |     0.3 |   +0%  ||  3214.46 |  3213.98 |   -0%  |  0.8134 |
 | TPC-H 07 ||      3.4 |     3.5 |   +1%  ||   291.13 |   287.71 |   -1%  |  0.5045 |
 | TPC-H 08 ||     26.0 |    25.6 |   -2%  ||    38.45 |    39.07 |   +2%  |  0.0927 |
 | TPC-H 09 ||      4.4 |     4.4 |   -1%  ||   227.63 |   228.65 |   +0%  |  0.2429 |
 | TPC-H 10 ||      1.7 |     1.7 |   -1%  ||   591.65 |   595.21 |   +1%  |  0.0000 |
…