@mrks 2020-11-20
> Partial changes from #2254 to simplify the review process.

Haven't looked at anything yet, but thank you!