@julianmenzler 2020-11-20
Alright, problem isolated. This PR's changes are responsible for the TPC-H Q20 performance drop.